Remove redundant decoration casts in 5 DLLs (tricky/effect2/effect4/gameui; oracle-verified, match% unchanged)

+++ src/main/dll/dll_001B_effect2.c
@@ -158,7 +158,7 @@ static ModgfxVertexData* modgfx_getActiveVertexBuffer(ModgfxState* state)
 
 static ModgfxVertexData* modgfx_getInactiveVertexBuffer(ModgfxState* state)
 {
-    return state->vertexBuffers[1 - (uint)state->activeVertexBufferIndex];
+    return state->vertexBuffers[1 - state->activeVertexBufferIndex];
 }
 
 static ModgfxActiveEffect** modgfx_getActiveEffectRegistry(void)
@@ -286,7 +286,7 @@ void modgfx_initExpgfxSpawnConfig(undefined4 param_1, undefined4 param_2, undefi
 
     setupWord = FUN_80286840();
     FUN_800033a8((int)&gExpgfxSpawnConfig, 0, EXPGFX_SPAWN_CONFIG_PREFIX_BYTES);
-    gExpgfxSpawnConfig.colorByte0.value = (u8)setupValue;
+    gExpgfxSpawnConfig.colorByte0.value = setupValue;
     gExpgfxSpawnConfig.behaviorFlags = setupValue & 0xff;
     gExpgfxSpawnConfig.velocityZ = lbl_803E00B0;
     gExpgfxSpawnConfig.startPosX.value = lbl_803E00B0;
@@ -297,7 +297,7 @@ void modgfx_initExpgfxSpawnConfig(undefined4 param_1, undefined4 param_2, undefi
     gExpgfxSpawnConfig.startPosZ.value = lbl_803E00B4;
     gExpgfxSpawnConfig.colorByte1.value = 0;
     gExpgfxSpawnConfig.colorByte1.lowByte = 0;
-    gExpgfxSpawnConfig.quadVertex3Pad06 = (s32)setupWord;
+    gExpgfxSpawnConfig.quadVertex3Pad06 = setupWord;
     *(undefined4*)&gExpgfxSpawnConfig.scale = scaleBits;
     gExpgfxSpawnConfig.texture.word = textureWord;
     gExpgfxSpawnConfig.colorByte0.lowByte = colorLowByte;
@@ -352,7 +352,7 @@ void modgfx_scrollVertexTexcoords(int stateArg, int command)
     activeVertexData = modgfx_getActiveVertexBuffer(state);
     for (i = 0; i < state->vertexCount; i++)
     {
-        if (wrapCountS == (int)state->vertexCount)
+        if (wrapCountS == state->vertexCount)
         {
             coord = activeVertexData->texCoordS;
             if (coord < 0x101)
@@ -364,7 +364,7 @@ void modgfx_scrollVertexTexcoords(int stateArg, int command)
                 activeVertexData->texCoordS = coord + -0x100;
             }
         }
-        if (wrapCountT == (int)state->vertexCount)
+        if (wrapCountT == state->vertexCount)
         {
             coord = activeVertexData->texCoordT;
             if (coord < 0x101)
@@ -577,15 +577,15 @@ void modgfx_updateEffectPosition(int stateArg, int command, int mode)
         {
             state->posStepX =
                 *(float*)(command + 4) /
-                (float)((double)CONCAT44(0x43300000, (int)state->blendFrameCount ^ 0x80000000) -
+                (float)((double)CONCAT44(0x43300000, state->blendFrameCount ^ 0x80000000) -
                     DOUBLE_803e00c8);
             state->posStepY =
                 ((ModgfxVertexGroupCmd*)command)->valueY /
-                (float)((double)CONCAT44(0x43300000, (int)state->blendFrameCount ^ 0x80000000) - biasS
+                (float)((double)CONCAT44(0x43300000, state->blendFrameCount ^ 0x80000000) - biasS
                 );
             state->posStepZ =
                 ((ModgfxVertexGroupCmd*)command)->valueZ /
-                (float)((double)CONCAT44(0x43300000, (int)state->blendFrameCount ^ 0x80000000) - biasS
+                (float)((double)CONCAT44(0x43300000, state->blendFrameCount ^ 0x80000000) - biasS
                 );
         }
         state->posCurX = state->posCurX + state->posStepX;
@@ -629,13 +629,13 @@ void modgfx_updateEffectRotation(int stateArg, int command, int mode)
         else
         {
             state->rotStepZ =
-                (short)(((int)targetRotZ - (int)state->rotOffsetZ) / (int)state->blendFrameCount
+                (short)(((int)targetRotZ - state->rotOffsetZ) / state->blendFrameCount
                 );
             state->rotStepY =
-                (short)(((int)targetRotY - (int)state->rotOffsetY) / (int)state->blendFrameCount
+                (short)(((int)targetRotY - state->rotOffsetY) / state->blendFrameCount
                 );
             state->rotStepX =
-                (short)(((int)targetRotX - (int)state->rotOffsetX) / (int)state->blendFrameCount
+                (short)(((int)targetRotX - state->rotOffsetX) / state->blendFrameCount
                 );
         }
     }
@@ -1523,7 +1523,7 @@ int Effect2_func04(void* sourceObj, int effectId, PartFxSpawnParams* spawnParams
     }
     cfg.behaviorFlags = 0;
     cfg.renderFlags = 0;
-    cfg.effectIdByte = (u8)effectId;
+    cfg.effectIdByte = effectId;
     cfg.attachedSource = sourceObj;
     cfg.startPosX = lbl_803DF884;
     cfg.startPosY = lbl_803DF884;
@@ -1566,20 +1566,20 @@ int Effect2_func04(void* sourceObj, int effectId, PartFxSpawnParams* spawnParams
         break;
     case 0x2b1:
         cfg.velocityX = lbl_80310560.vel[0][0] * (f32)(s32)
-        randomGetRange((s32)lbl_80310560.vel[0][1], (s32)lbl_80310560.vel[0][2]);
+        randomGetRange((s32)lbl_80310560.vel[0][1], lbl_80310560.vel[0][2]);
         cfg.velocityY = lbl_80310560.vel[1][0] * (f32)(s32)
-        randomGetRange((s32)lbl_80310560.vel[1][1], (s32)lbl_80310560.vel[1][2]);
+        randomGetRange((s32)lbl_80310560.vel[1][1], lbl_80310560.vel[1][2]);
         cfg.velocityZ = lbl_80310560.vel[2][0] * (f32)(s32)
-        randomGetRange((s32)lbl_80310560.vel[2][1], (s32)lbl_80310560.vel[2][2]);
+        randomGetRange((s32)lbl_80310560.vel[2][1], lbl_80310560.vel[2][2]);
         cfg.startPosX = lbl_80310560.vel[3][0] * (f32)(s32)
-        randomGetRange((s32)lbl_80310560.vel[3][1], (s32)lbl_80310560.vel[3][2]);
+        randomGetRange((s32)lbl_80310560.vel[3][1], lbl_80310560.vel[3][2]);
         cfg.startPosY = lbl_80310560.vel[4][0] * (f32)(s32)
-        randomGetRange((s32)lbl_80310560.vel[4][1], (s32)lbl_80310560.vel[4][2]);
+        randomGetRange((s32)lbl_80310560.vel[4][1], lbl_80310560.vel[4][2]);
         cfg.startPosZ = lbl_80310560.vel[5][0] * (f32)(s32)
-        randomGetRange((s32)lbl_80310560.vel[5][1], (s32)lbl_80310560.vel[5][2]);
+        randomGetRange((s32)lbl_80310560.vel[5][1], lbl_80310560.vel[5][2]);
         cfg.scale = lbl_80310560.vel[6][0] * (f32)(s32)
-        randomGetRange((s32)lbl_80310560.vel[6][1], (s32)lbl_80310560.vel[6][2]);
-        cfg.lifetimeFrames = randomGetRange((s32)lbl_80310560.g08[1], (s32)lbl_80310560.g08[2]) + (s32)lbl_80310560.g08[
+        randomGetRange((s32)lbl_80310560.vel[6][1], lbl_80310560.vel[6][2]);
+        cfg.lifetimeFrames = randomGetRange((s32)lbl_80310560.g08[1], lbl_80310560.g08[2]) + (s32)lbl_80310560.g08[
             0];
         cfg.colorWord0 = lbl_80310560.col[0];
         cfg.colorWord1 = lbl_80310560.col[1];
@@ -2125,8 +2125,8 @@ int Effect2_func04(void* sourceObj, int effectId, PartFxSpawnParams* spawnParams
         cfg.initialAlpha = 0xff;
         cfg.linkGroup = 0x10;
         cfg.behaviorFlags = 0x80480110;
-        if (effectId == 0x278) cfg.textureId = (s16)lbl_80310660[3];
-        else cfg.textureId = (s16)lbl_80310660[effectId - 0x260];
+        if (effectId == 0x278) cfg.textureId = lbl_80310660[3];
+        else cfg.textureId = lbl_80310660[effectId - 0x260];
         break;
     case 0x263:
     case 0x264:
@@ -2145,8 +2145,8 @@ int Effect2_func04(void* sourceObj, int effectId, PartFxSpawnParams* spawnParams
         cfg.initialAlpha = 0xff;
         cfg.linkGroup = 0x10;
         cfg.behaviorFlags = 0x480110;
-        if (effectId == 0x276) cfg.textureId = (s16)lbl_80310660[3];
-        else cfg.textureId = (s16)lbl_80310660[effectId - 0x263];
+        if (effectId == 0x276) cfg.textureId = lbl_80310660[3];
+        else cfg.textureId = lbl_80310660[effectId - 0x263];
         break;
     case 0x266:
     case 0x267:
@@ -2165,8 +2165,8 @@ int Effect2_func04(void* sourceObj, int effectId, PartFxSpawnParams* spawnParams
         cfg.initialAlpha = 0xff;
         cfg.linkGroup = 0x10;
         cfg.behaviorFlags = 0x480100;
-        if (effectId == 0x277) cfg.textureId = (s16)lbl_80310660[3];
-        else cfg.textureId = (s16)lbl_80310660[effectId - 0x266];
+        if (effectId == 0x277) cfg.textureId = lbl_80310660[3];
+        else cfg.textureId = lbl_80310660[effectId - 0x266];
         break;
     case 0x26d:
         cfg.startPosX = (f32)(s32)
